convert BarChartSparkline to ES6 module
